app/base/temp-buf.c (temp_buf_get_memsize) Add some casts to gsize for
2004-02-13  Manish Singh  <yosh@gimp.org>

        * app/base/temp-buf.c (temp_buf_get_memsize)
        * core/gimpviewable.c (gimp_viewable_get_memsize): Add some casts
        to gsize for ints to prevent overflows.

        * app/base/tile-manager.c (tile_manager_get_memsize): Same as above,
        except cast to gint64.

+++ app/base/temp-buf.c
@@ -569,9 +569,9 @@ temp_buf_get_memsize (TempBuf *temp_buf)
     }
   else
     {
-      memsize += (temp_buf->bytes *
-                  temp_buf->width *
-                  temp_buf->height);
+      memsize += ((gsize) temp_buf->bytes *
+                          temp_buf->width *
+                          temp_buf->height);
     }
 
   return memsize;
